Extracted MRI3DSubvolumeSegmentationDataset Class
